+++ x-pack/plugins/spaces/server/lib/saved_objects_client/lib/query_filters.js
